Split TileMetadata into two smaller classes
